fix: migrate

+++ app/database/migrations/2023_09_08_010330_create_payers.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payers', function (Blueprint $table) {
             $table->id();
             $table->uuid('uuid');
-            $table->integer("company_id");
+            $table->bigInteger("company_id");
             $table->string('first_name', 20);
             $table->string('last_name', 30);
             $table->string('email', 100);
